% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_gnn_results.R
\name{analyse_experts}
\alias{analyse_experts}
\title{Analyse and Visualise Expert Network Specialisation}
\usage{
analyse_experts(
  gnn_results,
  prepared_data,
  group_mappings,
  top_n_features = 10
)
}
\arguments{
\item{gnn_results}{A list object from \code{train_gnn()}.}

\item{prepared_data}{A list object from \code{prepare_data()}.}

\item{group_mappings}{A named list mapping numeric codes to labels.}

\item{top_n_features}{The number of top features to visualise.}
}
\value{
A list containing the importance tables and plots.
}
\description{
This function analyses the input weights of the expert sub-networks to
determine which features are most important for each subgroup. It supports
datasets with two or more subgroups.
}
