% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Create a Sankey Plot to Visualise Patient Routing}
\usage{
plot_sankey(raw_data, gnn_results, expert_results, group_mappings, group_var)
}
\arguments{
\item{raw_data}{The original, unscaled dataframe.}

\item{gnn_results}{The results object from \code{train_gnn()}.}

\item{expert_results}{The results object from \code{analyse_experts()}.}

\item{group_mappings}{A named list mapping numeric codes to labels (e.g., \code{list('0' = "Male", '1' = "Female")}).}

\item{group_var}{A string with the column name of the sensitive attribute in the raw_data.}
}
\value{
A ggplot object representing the Sankey diagram.
}
\description{
Generates a Sankey diagram showing how patients from different subgroups
are routed to different experts based on their learned feature profiles.
}
